chore(balance): refatoring assetBalance
